+++ addon-gui/src/components/AddonFolderCollapse.vue
