+++ addon-gui/src/components/AddonFolderDeleteModal.vue
