+++ addon-gui/src/components/AddonRepoCloneModal.vue
