+++ addon-gui/src/components/AddonRepoDeleteModal.vue
